==> 4.modélisation_autogluon.py <==
# -*- coding: utf-8 -*-
"""Modélisation_AutoGluon.ipynb

Automatically generated by Colab.

Original file is located at
    https://colab.research.google.com/drive/1hugUZzL1qr4G0g5ln3TJmisYJl1AqxBN

#Modélisation avec AutoGluon
"""

import pandas as pd
from autogluon.tabular import TabularPredictor

"""Chargement des donnees apres Feature Engineering"""

dossier_donnees = "./Data/"
chemin_train = dossier_donnees + "train_feature_engineering.csv"
chemin_test = dossier_donnees + "test_feature_engineering.csv"

train = pd.read_csv(chemin_train)
test = pd.read_csv(chemin_test)

"""Séparer les caractéristiques et la cible"""

cible = 'Exited'
caracteristiques = [col for col in train.columns if col != cible]

"""Initialisation et entraînement du modèle AutoGluon"""

# Ici, nous utilisons AutoGluon pour construire un modèle performant automatiquement.
model_path = "./autogluon_models/"
predictor = TabularPredictor(label=cible, path=model_path, eval_metric='roc_auc')
predictor.fit(train_data=train[caracteristiques + [cible]],
              time_limit=600,  # Temps limite en secondes
              presets='best_quality')  # Configuration pour une qualité maximale

"""Sauvegarder les prédictions sur le jeu de test"""

test['Predictions'] = predictor.predict_proba(test[caracteristiques])[:, 1]
test[['id', 'Predictions']].to_csv(dossier_donnees + "test_predictions_autogluon.csv", index=False)

"""Afficher les résultats d'évaluation"""

performance = predictor.leaderboard(train, silent=True)
print(performance)
