==> 7.modèle_ensemble.py <==
# -*- coding: utf-8 -*-
"""Modèle_ensemble.ipynb

Automatically generated by Colab.

Original file is located at
    https://colab.research.google.com/drive/1G9a4cgkJZaSRqOObb8nvHPfHG_ufRPSP

#Modèle d'ensemble
"""

import pandas as pd

"""Chargement des prédictions des différents modèles"""

dossier_donnees = "./Data/"
chemin_predictions_autogluon = dossier_donnees + "test_predictions_autogluon.csv"
chemin_predictions_catboost = dossier_donnees + "test_predictions_catboost.csv"

predictions_autogluon = pd.read_csv(chemin_predictions_autogluon)
predictions_catboost = pd.read_csv(chemin_predictions_catboost)

"""Combinaison des prédictions"""

# Ici, nous faisons une moyenne simple des prédictions pour créer un modèle ensemble.
ensemble_predictions = (
    predictions_autogluon['Predictions'] + predictions_catboost['Predictions_CatBoost']
) / 2

"""Ajout des prédictions combinées au DataFrame final"""

resultats_ensemble = predictions_autogluon[['id']].copy()
resultats_ensemble['Predictions_Ensemble'] = ensemble_predictions

"""Sauvegarde des résultats"""

resultats_ensemble.to_csv(dossier_donnees + "test_predictions_ensemble.csv", index=False)

"""Affichage des premières lignes pour vérification"""

print(resultats_ensemble.head())
